split kernel up so we can write unit tests
