--- .cargo/config.toml
@@ -1,4 +1,0 @@
-[target.aarch64-unknown-none]
-
-[build]
-target = "aarch64-unknown-none"
