+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["kernel"]
+members = ["kernel", "kernel_core"]
 resolver = "2"
