+++ justfile
@@ -5,6 +5,8 @@ out_dir := env("EH_OUTPUT_DIR", absolute_path("./.build"))
 img_dir := out_dir / "image"
 vendor_tool_dir := out_dir / "vendor"
 
+host_target_triple := `rustc --version --verbose | grep "host" | awk '{print $2}'`
+
 # Choose a task to run.
 default:
     @just --choose
@@ -13,15 +15,19 @@ default:
 clean:
     rm -rf {{out_dir}}
 
-# Build Rust crates.
-build cargo_args="":
-    cargo build {{ if build_profile == "release" { "--release" } else { "" } }} {{cargo_args}}
-
 # Check formatting, types and lints.
 check cargo_args="" clippy_args="":
     cargo fmt --check {{cargo_args}}
     cargo check --all-features {{cargo_args}}
     cargo clippy --all-features {{cargo_args}} -- -Dclippy::all -Wclippy::pedantic {{clippy_args}}
+
+# Test Rust crates that are testable on the host.
+test cargo_args="":
+    cargo test -p kernel_core --target {{host_target_triple}} {{cargo_args}}
+
+# Build Rust crates.
+build cargo_args="":
+    cargo build {{ if build_profile == "release" { "--release" } else { "" } }} --target aarch64-unknown-none {{cargo_args}}
 
 mkimage_bin := vendor_tool_dir / "u-boot/tools/mkimage"
