+++ kernel_core/Cargo.toml
@@ -1,10 +1,8 @@
 [package]
-name = "kernel"
+name = "kernel_core"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-kernel_core = { path = "../kernel_core"}
 itertools = { version = "^0.13", default-features = false }
 byteorder = { version = "^1.5", default-features = false }
-
