+++ kernel_core/src/platform/mod.rs
@@ -1,4 +1,3 @@
 //! Definitions and drivers for the ARM platform.
 
-pub mod bss;
 pub mod device_tree;
